feat: adopt proper folder structure
