+++ .github/workflows/actions.yaml
@@ -10,7 +10,7 @@ jobs:
           go-version: '1.18.2'
       - run: go version
       - run: go mod download
-      - run: go build
+      - run: go build cmd/cosmos-proposals-checker.go
   go-vet:
     runs-on: ubuntu-latest
     steps:
@@ -40,7 +40,7 @@ jobs:
       - name: Check out code into the Go module directory
         uses: actions/checkout@v2
       - name: Run tests
-        run: go test -coverprofile coverage.txt -v *.go
+        run: go test -coverprofile coverage.txt -v ./...
       - name: Download uploader
         run: curl -Os https://uploader.codecov.io/latest/linux/codecov
       - name: Make uploader executable
