+++ cmd/cosmos-proposals-checker.go
@@ -1,0 +1,33 @@
+package main
+
+import (
+	"github.com/spf13/cobra"
+	"main/pkg"
+	"main/pkg/logger"
+)
+
+func Execute(configPath string) {
+	app := pkg.NewApp(configPath)
+	app.Start()
+}
+
+func main() {
+	var ConfigPath string
+
+	rootCmd := &cobra.Command{
+		Use:  "cosmos-proposals-checker",
+		Long: "Checks the specific wallets on different chains for proposal votes.",
+		Run: func(cmd *cobra.Command, args []string) {
+			Execute(ConfigPath)
+		},
+	}
+
+	rootCmd.PersistentFlags().StringVar(&ConfigPath, "config", "", "Config file path")
+	if err := rootCmd.MarkPersistentFlagRequired("config"); err != nil {
+		logger.GetDefaultLogger().Fatal().Err(err).Msg("Could not set flag as required")
+	}
+
+	if err := rootCmd.Execute(); err != nil {
+		logger.GetDefaultLogger().Fatal().Err(err).Msg("Could not start application")
+	}
+}
