+++ templates/telegram/help.html
@@ -1,6 +1,7 @@
 <strong>cosmos-proposals-checker</strong>
 Notifies you about the proposals your wallets hasn't voted upon.
 Can understand the following commands:
+- /proposals - displays active proposals and your wallets' votes on them
 - /proposals_mute &lt;duration&gt; &lt;chain&gt; &lt;proposal ID&gt; - mute notifications for a specific proposal
 - /proposals_mutes - display the active proposals mutes list
 - /help - display this command
