+++ templates/templates.go
@@ -1,0 +1,6 @@
+package templates
+
+import "embed"
+
+//go:embed *
+var TemplatesFs embed.FS
