--- utils.go
@@ -1,11 +1,0 @@
-package main
-
-func Filter[T any](slice []T, f func(T) bool) []T {
-	var n []T
-	for _, e := range slice {
-		if f(e) {
-			n = append(n, e)
-		}
-	}
-	return n
-}
